hostconfig: move unit tests to api/types/container
These types were moved to api/types/container in 7ac4232e70fe7cf7318333cd0890db7f95663079,
but the unit-tests for them were not moved. This patch moves the unit-tests back together
with the types.
